Them Thong Bao Department, Employee

+++ src/app/pages/departments/list/list.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import {EmployeeModel} from "../../../@core/interfaces/employee.interface";
 import {DepartmentModel} from "../../../@core/interfaces/department.interface";
 import {DepartmentService} from "../../../@core/services/apis/department.service";
+import { DeleteComponent } from '../delete/delete.component';
+import { NbDialogService, NbToastrService } from '@nebular/theme';
 
 @Component({
   selector: 'app-list',
@@ -11,7 +13,9 @@ import {DepartmentService} from "../../../@core/services/apis/department.service
 export class ListComponent implements OnInit{
   departments!: DepartmentModel[]
   constructor(
-    private DepartmentService: DepartmentService) {
+    private DepartmentService: DepartmentService,
+    private dialogService: NbDialogService,
+    private toastrService: NbToastrService,) {
     this.DepartmentService.getAllDepartment().subscribe({
       next: (data) => {
         this.departments = data;
@@ -32,13 +36,30 @@ export class ListComponent implements OnInit{
     })
   }
 
+  // delete(id: DepartmentModel) {
+  //   this.DepartmentService.deleteDepartment(id).subscribe({
+  //     next: () => {
+  //       console.log('xoa thanh cong');
+  //     },
+  //     error: (error) => {
+  //       console.log(error);
+  //     }
+  //   });
+  // }
+
   delete(id: DepartmentModel) {
-    this.DepartmentService.deleteDepartment(id).subscribe({
-      next: () => {
-        console.log('xoa thanh cong');
-      },
-      error: (error) => {
-        console.log(error);
+    this.dialogService.open(DeleteComponent)
+      .onClose.subscribe((confirmed: boolean) => {
+      if (confirmed) {
+        this.DepartmentService.deleteDepartment(id).subscribe({
+          next: res => {
+            this.toastrService.show('Xóa phòng ban thành công!', 'Thành công', { status: 'success' });
+            this.getAllDepartment();
+          },
+          error: err => {
+            this.toastrService.show('Xóa phòng ban thất bại. Vui lòng thử lại.', 'Lỗi', { status: 'danger' });
+          }
+        });
       }
     });
   }
